Correct entities creation and create repositories

+++ src/main/java/com/yuri/freire/Cash_Stream/entities/Token.java
@@ -8,8 +8,6 @@ import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
-import java.math.BigInteger;
-
 @EqualsAndHashCode(callSuper = true)
 @Data
 @SuperBuilder
@@ -30,7 +28,7 @@ public class Token extends BaseEntity{
             allocationSize = 1,
             initialValue = 1
     )
-    private BigInteger id;
+    private Integer id;
 
     private String token;
