+++ src/main/java/com/yuri/freire/Cash_Stream/entities/repositories/UserRespository.java
@@ -3,9 +3,8 @@ package com.yuri.freire.Cash_Stream.entities.repositories;
 import com.yuri.freire.Cash_Stream.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
-public interface UserRespository extends JpaRepository<User, BigInteger> {
+public interface UserRespository extends JpaRepository<User, Integer> {
     Optional<User> findByUsername(String username);
 }
